Fixed a bug for large files
When using sendfile(2) we cannot specify MSG_MORE to indicate that the
hash should not be fully computed. Therfore, we calculated only around
the first 2 GiB of every file. To indicate MSG_MORE but still be able to
use in-kernel copying without going through userspace, we can use
splice(2). However, splice(2) requires one side to be a pipe. Therefore,
we now create a pipe(2), splice(2) from the file descriptor into the
writing fd of the pipe and perform a second splice(2) with MSG_MORE from
the reading end of the pipe into the socket descriptor. Annoying and a
bit slower than before, but produces consistent results.